Fixes tumblr SVC authentication
- Fixes the svc service access authentication. Instead of querying a
different host (https://www.tumblr.com/blog_auth/) and extract a key
that is used in the following POST requests to validate the cookie, the
svc service now uses the same mechanism that is also used in the tumblr
search. A key has to be extracted from the main blog page
(https://www.tumblr.com/dashboard/blog/blogname) that is subsequently
added to the POST headers in each following request. The key depends on
the cookie and thus on your login (username).