Allows to load the Tumblr ToS cookie separately.
Previously, TumblThree agreed to the new Tumblr ToS (GDPR changes) at
each startup because there is an issue storing and loading cookies
without subdomain using the CookieContainer as well as the wininet.dll.
You need a URI to specify the cookies you want, but this fails is they
are prefixed with a dot for a whole domain cookie as it is the case for
the Tumblr ToS cookie.

That cookie interfered with the ToS cookie retrieved from logging into
Tumblr, hence invalidating the login.

The TumblrBlogDetector.cs now uses only the ToS cookie that is generated
at startup without login information to be able to differentiate between
blogs, hidden blogs, tumblr searches or tumblr liked-by downloads that
either require a login or can be access without being logged in.